Move the generate functionality into the main task.
In some cases, the generate task won't properly detect that the makepot
task has run and will stop the current process. There's no need for the
secondary task at the moment, so moving that functionality into the main
task will remove the depedency and the need to pass around a global
option.