Add checktextdomain task
This is a fist pass.

This is a fork from  https://github.com/stephenharris/grunt-checktextdomain and
the majority of the commit is from code Stephen Harris wrote.

This commit:

* Changes option names to be camel case instead of underscores, as per common
practice.
* Allows createReportFile to accept a file path string instead of just being a
boolean. This is used within the unit tests to keep the generated report files
within the tmp/ directory. This directory is cleaned before testing, so it
avoids having to tidy up each file during the testing as it was in the original
code.
* Passes all existing unit tests. No functional testing done yet.
* Merges existing documentation into README.md.